Test `ArrayableTrait` separately
