+++ tests/Driver/Mysql/ArrayableTraitTest.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\ActiveRecord\Tests\Driver\Mysql;
+
+use Yiisoft\ActiveRecord\Tests\Support\MysqlHelper;
+use Yiisoft\Db\Connection\ConnectionInterface;
+
+final class ArrayableTraitTest extends \Yiisoft\ActiveRecord\Tests\ArrayableTraitTest
+{
+    protected function createConnection(): ConnectionInterface
+    {
+        return (new MysqlHelper())->createConnection();
+    }
+}
