+++ tests/MagicActiveRecordTest.php
@@ -11,8 +11,6 @@ use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\Alpha;
 use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\Animal;
 use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\Cat;
 use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\Customer;
-use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\CustomerClosureField;
-use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\CustomerForArrayable;
 use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\CustomerWithAlias;
 use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\CustomerWithProperties;
 use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord\Dog;
@@ -734,102 +732,6 @@ abstract class MagicActiveRecordTest extends TestCase
         }
 
         $this->assertEquals($eagerItemsCount, $lazyItemsCount);
-    }
-
-    public function testToArray(): void
-    {
-        $this->checkFixture($this->db(), 'customer', true);
-
-        $customerQuery = new ActiveQuery(Customer::class);
-        $customer = $customerQuery->findOne(1);
-
-        $this->assertSame(
-            [
-                'id' => 1,
-                'email' => 'user1@example.com',
-                'name' => 'user1',
-                'address' => 'address1',
-                'status' => 1,
-                'bool_status' => true,
-                'profile_id' => 1,
-            ],
-            $customer->toArray(),
-        );
-    }
-
-    public function testToArrayWithClosure(): void
-    {
-        $this->checkFixture($this->db(), 'customer', true);
-
-        $customerQuery = new ActiveQuery(CustomerClosureField::class);
-        $customer = $customerQuery->findOne(1);
-
-        $this->assertSame(
-            [
-                'id' => 1,
-                'email' => 'user1@example.com',
-                'name' => 'user1',
-                'address' => 'address1',
-                'status' => 'active',
-                'bool_status' => true,
-                'profile_id' => 1,
-            ],
-            $customer->toArray(),
-        );
-    }
-
-    public function testToArrayForArrayable(): void
-    {
-        $this->checkFixture($this->db(), 'customer', true);
-
-        $customerQuery = new ActiveQuery(CustomerForArrayable::class);
-
-        /** @var CustomerForArrayable $customer */
-        $customer = $customerQuery->findOne(1);
-        /** @var CustomerForArrayable $customer2 */
-        $customer2 = $customerQuery->findOne(2);
-        /** @var CustomerForArrayable $customer3 */
-        $customer3 = $customerQuery->findOne(3);
-
-        $customer->setItem($customer2);
-        $customer->setItems($customer3);
-
-        $this->assertSame(
-            [
-                'id' => 1,
-                'email' => 'user1@example.com',
-                'name' => 'user1',
-                'address' => 'address1',
-                'status' => 'active',
-                'item' => [
-                    'id' => 2,
-                    'email' => 'user2@example.com',
-                    'name' => 'user2',
-                    'status' => 'active',
-                ],
-                'items' => [
-                    [
-                        'id' => 3,
-                        'email' => 'user3@example.com',
-                        'name' => 'user3',
-                        'status' => 'inactive',
-                    ],
-                ],
-            ],
-            $customer->toArray([
-                'id',
-                'name',
-                'email',
-                'address',
-                'status',
-                'item.id',
-                'item.name',
-                'item.email',
-                'items.0.id',
-                'items.0.name',
-                'items.0.email',
-            ]),
-        );
     }
 
     public function testSaveWithoutChanges(): void
