--- tests/Stubs/MagicActiveRecord/CustomerClosureField.php
@@ -1,33 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord;
-
-use Yiisoft\ActiveRecord\Tests\Stubs\MagicActiveRecord;
-
-/**
- * Class CustomerClosureField.
- *
- * @property int $id
- * @property string $name
- * @property string $email
- * @property string $address
- * @property int $status
- */
-final class CustomerClosureField extends MagicActiveRecord
-{
-    public function getTableName(): string
-    {
-        return 'customer';
-    }
-
-    public function fields(): array
-    {
-        $fields = parent::fields();
-
-        $fields['status'] = static fn(self $customer) => $customer->status === 1 ? 'active' : 'inactive';
-
-        return $fields;
-    }
-}
